Remove ConfigurableEndpoint
