+++ src/NServiceBus.Core/ConfiguredEndpoint.cs
@@ -1,28 +1,137 @@
 namespace NServiceBus
 {
     using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
     using System.Threading.Tasks;
     using Features;
+    using Hosting.Helpers;
     using Installation;
+    using MessageInterfaces;
+    using MessageInterfaces.MessageMapper.Reflection;
     using ObjectBuilder;
+    using Routing;
+    using Routing.MessageDrivenSubscriptions;
     using Settings;
     using Transport;
+    using Unicast.Messages;
 
-    abstract class ConfiguredEndpoint
+    class ConfiguredEndpoint
     {
-        protected ConfiguredEndpoint(ReceiveComponent receiveComponent, QueueBindings queueBindings, FeatureActivator featureActivator, TransportInfrastructure transportInfrastructure, CriticalError criticalError, SettingsHolder settings, PipelineComponent pipelineComponent, ContainerComponent containerComponent)
+        public ConfiguredEndpoint(SettingsHolder settings,
+            ContainerComponent containerComponent,
+            PipelineComponent pipelineComponent)
         {
-            this.receiveComponent = receiveComponent;
-            this.queueBindings = queueBindings;
-            this.featureActivator = featureActivator;
-            this.transportInfrastructure = transportInfrastructure;
-            this.criticalError = criticalError;
             this.settings = settings;
-            this.pipelineComponent = pipelineComponent;
             this.containerComponent = containerComponent;
+            this.pipelineComponent = pipelineComponent;
+        }
+        static void ScanTypesAndDoOtherStuff(EndpointConfiguration endpointConfiguration)
+        {
+            var scannedTypes = endpointConfiguration.ScannedTypes;
+
+            if (scannedTypes == null)
+            {
+                var directoryToScan = AppDomain.CurrentDomain.RelativeSearchPath ?? AppDomain.CurrentDomain.BaseDirectory;
+
+                scannedTypes = GetAllowedTypes(directoryToScan, endpointConfiguration.Settings);
+            }
+            else
+            {
+                scannedTypes = scannedTypes.Union(GetAllowedCoreTypes(endpointConfiguration.Settings)).ToList();
+            }
+
+            endpointConfiguration.Settings.SetDefault("TypesToScan", scannedTypes);
+            ActivateAndInvoke<INeedInitialization>(scannedTypes, t => t.Customize(endpointConfiguration));
+
+            var conventions = endpointConfiguration.ConventionsBuilder.Conventions;
+            endpointConfiguration.Settings.SetDefault(conventions);
+
+            ConfigureMessageTypes(conventions, endpointConfiguration.Settings);
         }
 
-        public async Task<IStartableEndpoint> Initialize()
+        public static IConfiguredEndpointWithExternalContainer ConfigureWithExternalContainer(EndpointConfiguration endpointConfiguration, IConfigureComponents configureComponents)
+        {
+            ScanTypesAndDoOtherStuff(endpointConfiguration);
+
+            endpointConfiguration.ContainerComponent.InitializeWithExternalContainer(configureComponents);
+
+            var configured = new ConfiguredExternalContainerEndpoint(endpointConfiguration.Settings, endpointConfiguration.ContainerComponent, endpointConfiguration.PipelineComponent);
+            configured.Initialize();
+
+            return configured;
+
+        }
+
+        public static ConfiguredInternalContainerEndpoint ConfigureWithInternalContainer(EndpointConfiguration endpointConfiguration)
+        {
+            ScanTypesAndDoOtherStuff(endpointConfiguration);
+
+            endpointConfiguration.ContainerComponent.InitializeWithInternalContainer();
+
+            var configured = new ConfiguredInternalContainerEndpoint(endpointConfiguration.Settings, endpointConfiguration.ContainerComponent, endpointConfiguration.PipelineComponent);
+            configured.Initialize();
+
+            return configured;
+        }
+
+        void Initialize()
+        {
+            containerComponent.ContainerConfiguration.RegisterSingleton<ReadOnlySettings>(settings);
+
+            RegisterCriticalErrorHandler();
+
+            var concreteTypes = settings.GetAvailableTypes()
+                .Where(IsConcrete)
+                .ToList();
+
+            featureActivator = BuildFeatureActivator(concreteTypes);
+
+            ConfigRunBeforeIsFinalized(concreteTypes);
+
+            transportInfrastructure = InitializeTransportComponent();
+
+            var receiveConfiguration = BuildReceiveConfiguration(transportInfrastructure);
+
+            var routing = InitializeRouting(transportInfrastructure, receiveConfiguration);
+
+            var messageMapper = new MessageMapper();
+            settings.Set<IMessageMapper>(messageMapper);
+
+            pipelineComponent.AddRootContextItem<IMessageMapper>(messageMapper);
+
+            var featureStats = featureActivator.SetupFeatures(containerComponent.ContainerConfiguration, pipelineComponent.PipelineSettings, routing, receiveConfiguration);
+            settings.AddStartupDiagnosticsSection("Features", featureStats);
+
+            pipelineComponent.RegisterBehaviorsInContainer(containerComponent.ContainerConfiguration);
+            containerComponent.ContainerConfiguration.ConfigureComponent(b => settings.Get<Notifications>(), DependencyLifecycle.SingleInstance);
+
+            var eventAggregator = new EventAggregator(settings.Get<NotificationSubscriptions>());
+
+            pipelineComponent.AddRootContextItem<IEventAggregator>(eventAggregator);
+
+            var shouldRunInstallers = settings.GetOrDefault<bool>("Installers.Enable");
+
+            if (shouldRunInstallers)
+            {
+                RegisterInstallers(concreteTypes);
+            }
+
+            settings.AddStartupDiagnosticsSection("Endpoint",
+                new
+                {
+                    Name = settings.EndpointName(),
+                    SendOnly = settings.Get<bool>("Endpoint.SendOnly"),
+                    NServiceBusVersion = GitVersionInformation.MajorMinorPatch
+                }
+            );
+
+            queueBindings = settings.Get<QueueBindings>();
+            receiveComponent = CreateReceiveComponent(receiveConfiguration, transportInfrastructure, pipelineComponent, queueBindings, eventAggregator);
+        }
+
+        public async Task<IStartableEndpoint> CreateStartableEndpoint()
         {
             pipelineComponent.Initialize(containerComponent.Builder);
 
@@ -40,7 +149,7 @@ namespace NServiceBus
                 await RunInstallers(containerComponent.Builder, username).ConfigureAwait(false);
             }
 
-            messageSession = new MessageSession(pipelineComponent.CreateRootContext(containerComponent.Builder));
+            var messageSession = new MessageSession(pipelineComponent.CreateRootContext(containerComponent.Builder));
 
             return new StartableEndpoint(settings, containerComponent, featureActivator, transportInfrastructure, receiveComponent, criticalError, messageSession);
         }
@@ -70,15 +179,245 @@ namespace NServiceBus
             return userName;
         }
 
-        MessageSession messageSession;
-        ReceiveComponent receiveComponent;
-        QueueBindings queueBindings;
+        RoutingComponent InitializeRouting(TransportInfrastructure transportInfrastructure, ReceiveConfiguration receiveConfiguration)
+        {
+            // use GetOrCreate to use of instances already created during EndpointConfiguration.
+            var routing = new RoutingComponent(
+                settings.GetOrCreate<UnicastRoutingTable>(),
+                settings.GetOrCreate<DistributionPolicy>(),
+                settings.GetOrCreate<EndpointInstances>(),
+                settings.GetOrCreate<Publishers>());
+
+            routing.Initialize(settings, transportInfrastructure.ToTransportAddress, pipelineComponent.PipelineSettings, receiveConfiguration);
+
+            return routing;
+        }
+
+        TransportInfrastructure InitializeTransportComponent()
+        {
+            if (!settings.HasExplicitValue<TransportDefinition>())
+            {
+                throw new Exception("A transport has not been configured. Use 'EndpointConfiguration.UseTransport()' to specify a transport.");
+            }
+
+            var transportDefinition = settings.Get<TransportDefinition>();
+            var connectionString = settings.Get<TransportConnectionString>().GetConnectionStringOrRaiseError(transportDefinition);
+            var transportInfrastructure = transportDefinition.Initialize(settings, connectionString);
+            settings.Set(transportInfrastructure);
+
+            var transportType = transportDefinition.GetType();
+
+            settings.AddStartupDiagnosticsSection("Transport", new
+            {
+                Type = transportType.FullName,
+                Version = FileVersionRetriever.GetFileVersion(transportType)
+            });
+
+            return transportInfrastructure;
+        }
+
+        ReceiveConfiguration BuildReceiveConfiguration(TransportInfrastructure transportInfrastructure)
+        {
+            var receiveConfiguration = ReceiveConfigurationBuilder.Build(settings, transportInfrastructure);
+
+            if (receiveConfiguration == null)
+            {
+                return null;
+            }
+
+            //note: remove once settings.LogicalAddress() , .LocalAddress() and .InstanceSpecificQueue() has been obsoleted
+            settings.Set(receiveConfiguration);
+
+            return receiveConfiguration;
+        }
+
+        ReceiveComponent CreateReceiveComponent(ReceiveConfiguration receiveConfiguration,
+            TransportInfrastructure transportInfrastructure,
+            PipelineComponent pipeline,
+            QueueBindings queueBindings,
+            EventAggregator eventAggregator)
+        {
+            var errorQueue = settings.ErrorQueueAddress();
+
+            var receiveComponent = new ReceiveComponent(receiveConfiguration,
+                receiveConfiguration != null ? transportInfrastructure.ConfigureReceiveInfrastructure() : null, //don't create the receive infrastructure for send-only endpoints
+                pipeline,
+                eventAggregator,
+                criticalError,
+                errorQueue);
+
+            receiveComponent.BindQueues(queueBindings);
+
+            if (receiveConfiguration != null)
+            {
+                settings.AddStartupDiagnosticsSection("Receiving", new
+                {
+                    receiveConfiguration.LocalAddress,
+                    receiveConfiguration.InstanceSpecificQueue,
+                    receiveConfiguration.LogicalAddress,
+                    receiveConfiguration.PurgeOnStartup,
+                    receiveConfiguration.QueueNameBase,
+                    TransactionMode = receiveConfiguration.TransactionMode.ToString("G"),
+                    receiveConfiguration.PushRuntimeSettings.MaxConcurrency,
+                    Satellites = receiveConfiguration.SatelliteDefinitions.Select(s => new
+                    {
+                        s.Name,
+                        s.ReceiveAddress,
+                        TransactionMode = s.RequiredTransportTransactionMode.ToString("G"),
+                        s.RuntimeSettings.MaxConcurrency
+                    }).ToArray()
+                });
+            }
+
+            return receiveComponent;
+        }
+
+        static bool IsConcrete(Type x)
+        {
+            return !x.IsAbstract && !x.IsInterface;
+        }
+
+        void ConfigRunBeforeIsFinalized(IEnumerable<Type> concreteTypes)
+        {
+            foreach (var instanceToInvoke in concreteTypes.Where(IsIWantToRunBeforeConfigurationIsFinalized)
+                .Select(type => (IWantToRunBeforeConfigurationIsFinalized)Activator.CreateInstance(type)))
+            {
+                instanceToInvoke.Run(settings);
+            }
+        }
+
+        static bool IsIWantToRunBeforeConfigurationIsFinalized(Type type)
+        {
+            return typeof(IWantToRunBeforeConfigurationIsFinalized).IsAssignableFrom(type);
+        }
+
+        FeatureActivator BuildFeatureActivator(IEnumerable<Type> concreteTypes)
+        {
+            var featureActivator = new FeatureActivator(settings);
+            foreach (var type in concreteTypes.Where(t => IsFeature(t)))
+            {
+                featureActivator.Add(type.Construct<Feature>());
+            }
+            return featureActivator;
+        }
+
+        static bool IsFeature(Type type)
+        {
+            return typeof(Feature).IsAssignableFrom(type);
+        }
+
+        void RegisterCriticalErrorHandler()
+        {
+            settings.TryGet("onCriticalErrorAction", out Func<ICriticalErrorContext, Task> errorAction);
+            criticalError = new CriticalError(errorAction);
+            containerComponent.ContainerConfiguration.RegisterSingleton(criticalError);
+        }
+
+        void RegisterInstallers(IEnumerable<Type> concreteTypes)
+        {
+            foreach (var installerType in concreteTypes.Where(t => IsINeedToInstallSomething(t)))
+            {
+                containerComponent.ContainerConfiguration.ConfigureComponent(installerType, DependencyLifecycle.InstancePerCall);
+            }
+        }
+
+        static List<Type> GetAllowedTypes(string path, SettingsHolder settings)
+        {
+            var assemblyScannerSettings = settings.GetOrCreate<AssemblyScannerConfiguration>();
+            var assemblyScanner = new AssemblyScanner(path)
+            {
+                AssembliesToSkip = assemblyScannerSettings.ExcludedAssemblies,
+                TypesToSkip = assemblyScannerSettings.ExcludedTypes,
+                ScanNestedDirectories = assemblyScannerSettings.ScanAssembliesInNestedDirectories,
+                ThrowExceptions = assemblyScannerSettings.ThrowExceptions,
+                ScanAppDomainAssemblies = assemblyScannerSettings.ScanAppDomainAssemblies
+            };
+
+            return Scan(assemblyScanner, settings);
+        }
+
+        static List<Type> GetAllowedCoreTypes(SettingsHolder settings)
+        {
+            var assemblyScannerSettings = settings.GetOrCreate<AssemblyScannerConfiguration>();
+            var assemblyScanner = new AssemblyScanner(Assembly.GetExecutingAssembly())
+            {
+                AssembliesToSkip = assemblyScannerSettings.ExcludedAssemblies,
+                TypesToSkip = assemblyScannerSettings.ExcludedTypes,
+                ScanNestedDirectories = assemblyScannerSettings.ScanAssembliesInNestedDirectories,
+                ThrowExceptions = assemblyScannerSettings.ThrowExceptions,
+                ScanAppDomainAssemblies = assemblyScannerSettings.ScanAppDomainAssemblies
+            };
+
+            return Scan(assemblyScanner, settings);
+        }
+
+        static List<Type> Scan(AssemblyScanner assemblyScanner, SettingsHolder settings)
+        {
+            var results = assemblyScanner.GetScannableAssemblies();
+
+            settings.AddStartupDiagnosticsSection("AssemblyScanning", new
+            {
+                Assemblies = results.Assemblies.Select(a => a.FullName),
+                results.ErrorsThrownDuringScanning,
+                results.SkippedFiles
+            });
+
+            return results.Types;
+        }
+
+        static void ActivateAndInvoke<T>(IList<Type> types, Action<T> action) where T : class
+        {
+            ForAllTypes<T>(types, t =>
+            {
+                if (!HasDefaultConstructor(t))
+                {
+                    throw new Exception($"Unable to create the type '{t.Name}'. Types implementing '{typeof(T).Name}' must have a public parameterless (default) constructor.");
+                }
+
+                var instanceToInvoke = (T)Activator.CreateInstance(t);
+                action(instanceToInvoke);
+            });
+        }
+
+        static void ForAllTypes<T>(IEnumerable<Type> types, Action<Type> action) where T : class
+        {
+            // ReSharper disable HeapView.SlowDelegateCreation
+            foreach (var type in types.Where(t => typeof(T).IsAssignableFrom(t) && !(t.IsAbstract || t.IsInterface)))
+            {
+                action(type);
+            }
+            // ReSharper restore HeapView.SlowDelegateCreation
+        }
+
+        static void ConfigureMessageTypes(Conventions conventions, SettingsHolder settings)
+        {
+            var messageMetadataRegistry = new MessageMetadataRegistry(conventions.IsMessageType);
+
+            messageMetadataRegistry.RegisterMessageTypesFoundIn(settings.GetAvailableTypes());
+
+            settings.Set(messageMetadataRegistry);
+
+            var foundMessages = messageMetadataRegistry.GetAllMessages().ToList();
+
+            settings.AddStartupDiagnosticsSection("Messages", new
+            {
+                CustomConventionUsed = conventions.CustomMessageTypeConventionUsed,
+                NumberOfMessagesFoundAtStartup = foundMessages.Count,
+                Messages = foundMessages.Select(m => m.MessageType.FullName)
+            });
+        }
+
+        static bool HasDefaultConstructor(Type type) => type.GetConstructor(Type.EmptyTypes) != null;
+
+        static bool IsINeedToInstallSomething(Type t) => typeof(INeedToInstallSomething).IsAssignableFrom(t);
+
+        PipelineComponent pipelineComponent;
+        SettingsHolder settings;
+        ContainerComponent containerComponent;
+        CriticalError criticalError;
         FeatureActivator featureActivator;
         TransportInfrastructure transportInfrastructure;
-        CriticalError criticalError;
-        SettingsHolder settings;
-        PipelineComponent pipelineComponent;
-        ContainerComponent containerComponent;
-      
+        QueueBindings queueBindings;
+        ReceiveComponent receiveComponent;
     }
 }
